[Android] Added water intake date grouping

+++ androidApp/src/main/kotlin/com/fitnest/android/screen/private_area/home/data/HomeScreenData.kt
@@ -35,7 +35,7 @@ data class HomeScreenData(
     )
 
     data class WaterIntakeSubWidget(
-        val amount: Int? = null,
+        val amount: Double? = null,
         val progress: Double? = null,
         val intakes: List<WaterIntake>? = null
     )
